feat :: 댓글 삭제 기능

+++ src/main/java/com/example/springboot_practice/domain/feed/application/DeleteFeedService.java
@@ -3,6 +3,7 @@ package com.example.springboot_practice.domain.feed.application;
 import com.example.springboot_practice.domain.feed.application.facade.FeedFacade;
 import com.example.springboot_practice.domain.feed.domain.Feed;
 import com.example.springboot_practice.domain.feed.domain.repository.FeedRepository;
+import com.example.springboot_practice.domain.feed.exception.CannotModifyFeedException;
 import com.example.springboot_practice.domain.user.application.facade.UserFacade;
 import com.example.springboot_practice.domain.user.domain.User;
 import lombok.RequiredArgsConstructor;
@@ -20,6 +21,10 @@ public class DeleteFeedService {
     public void deleteFeed(String title) {
         User user = userFacade.currentUser();
         Feed feed = feedFacade.getFeeds(title);
+
+        if (!user.equals(feed.getUser())) {
+            throw CannotModifyFeedException.EXCEPTION;
+        }
 
         feedRepository.deleteByTitle(title);
     }
